feat: polish hypergraph ui for sidebar compability
