+++ apps/torus-portal/src/app/(pages)/(permission-graph)/_components/my-agent-button.tsx
@@ -1,34 +1,29 @@
 "use client";
 
-import { User } from "lucide-react";
+import { Diameter } from "lucide-react";
 
 import { useTorus } from "@torus-ts/torus-provider";
 import { Button } from "@torus-ts/ui/components/button";
 
-import type {
-  CustomGraphData,
-  CustomGraphNode,
-} from "./permission-graph-types";
+import { useGraphData } from "./force-graph/use-graph-data";
+import type { CustomGraphNode } from "./permission-graph-types";
 
 interface MyAgentButtonProps {
-  graphData: CustomGraphData | null;
   onNodeClick: (node: CustomGraphNode) => void;
 }
 
-export function MyAgentButton({ graphData, onNodeClick }: MyAgentButtonProps) {
+export function MyAgentButton({ onNodeClick }: MyAgentButtonProps) {
   const { selectedAccount } = useTorus();
+  const { graphData } = useGraphData();
 
-  // Don't show button if not logged in or no graph data
   if (!selectedAccount?.address || !graphData) {
     return null;
   }
 
-  // Find if the logged-in user's address exists as a node in the graph
   const userNode = graphData.nodes.find(
     (node) => node.id.toLowerCase() === selectedAccount.address.toLowerCase(),
   );
 
-  // Don't show button if user doesn't have an agent node
   if (!userNode) {
     return null;
   }
@@ -40,12 +35,11 @@ export function MyAgentButton({ graphData, onNodeClick }: MyAgentButtonProps) {
   return (
     <Button
       variant="outline"
-      size="sm"
       onClick={handleClick}
-      className="h-9 animate-fade-down animate-delay-200 w-full md:w-fit"
+      className="h-[2.60rem] rounded animate-fade-up animate-delay-200"
     >
-      <User className="w-4 h-4 mr-2" />
-      View my agent
+      <Diameter className="w-4 h-4 mr-0.5" />
+      Open my agent
     </Button>
   );
 }
